I have to be very clear
I hate you gamers put it exclusive inside Discord server. WTF?!
you ban and disallow them to have it for good.
I know that too, burner account detection poops!

Pls!!! make it easy to download, and nicely
ask to credit, that's it! I'll be the good example
putting the sauce in the description
so hopefully other gamers
would do the same.

pecking hell!